first working implementation of Touch/Face ID on Share Extension
