+++ Wire-iOS/Sources/Components/Settings/SettingsProperty.swift
@@ -16,78 +16,6 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 // 
 
-
-
-/**
-Available settings
-
-- ChatHeadsDisabled:      Disable chat heads in conversation and self profile
-- DisableMarkdown:        Disable markdown formatter for messages
-- SkipFirstTimeUseChecks: Temporarily skip firts time checks
-- PreferredFlashMode:     Flash mode for internal camera UI
-- DarkMode:               Dark mode for conversation
-- PriofileName:           User name
-- SoundAlerts:            Sound alerts level
-- DisableCrashAndAnalyticsSharing: Opt-Out analytics and Hockey
-- DisableSendButton:      Opt-Out of new send button
-- DisableLinkPreviews:    Disable link previews for links you send
-- Disable(.*):            Disable some app features (debug)
-*/
-enum SettingsPropertyName: String, CustomStringConvertible {
-    
-    // User defaults
-    case chatHeadsDisabled = "ChatHeadsDisabled"
-    case notificationContentVisible = "NotificationContentVisible"
-    case disableMarkdown = "Markdown"
-    
-    case skipFirstTimeUseChecks = "SkipFirstTimeUseChecks"
-    
-    case preferredFlashMode = "PreferredFlashMode"
-    
-    case darkMode = "DarkMode"
-    
-    case disableSendButton = "DisableSendButton"
-    
-    case disableLinkPreviews = "DisableLinkPreviews"
-    
-    // Profile
-    case profileName = "ProfileName"
-    case accentColor = "AccentColor"
-    
-    // AVS
-    case soundAlerts = "SoundAlerts"
-    case callingConstantBitRate = "constantBitRate"
-    
-    // Analytics
-    case disableCrashAndAnalyticsSharing = "DisableCrashAndAnalyticsSharing"
-
-    // Sounds
-    case messageSoundName = "MessageSoundName"
-    case callSoundName = "CallSoundName"
-    case pingSoundName = "PingSoundName"
-
-    // Open In
-    case tweetOpeningOption = "TweetOpeningOption"
-    case mapsOpeningOption = "MapsOpeningOption"
-    case browserOpeningOption = "BrowserOpeningOption"
-    
-    // Debug
-    case disableCallKit = "DisableCallKit"
-    case callingProtocolStrategy = "CallingProtcolStrategy"
-    case enableBatchCollections = "EnableBatchCollections"
-
-    case lockApp = "lockApp"
-    case lockAppLastDate = "lockAppLastDate"
-        
-    var changeNotificationName: String {
-        return self.description + "ChangeNotification"
-    }
-    
-    var description: String {
-        return self.rawValue;
-    }
-}
-
 enum SettingsPropertyValue: Equatable {
     case number(value: NSNumber)
     case string(value: Swift.String)
